change default motd & moved to external proxy-plugin
