+++ addons/proxy-addon/build.gradle.kts
@@ -1,4 +1,6 @@
 dependencies {
+    implementation(project(":api"))
+
     implementation(libs.velocity)
     annotationProcessor(libs.velocity)
 }
