+++ addons/proxy-addon/src/main/java/dev/httpmarco/polocloud/proxy/VelocityPlatformPlugin.java
@@ -16,16 +16,36 @@
 
 package dev.httpmarco.polocloud.proxy;
 
+import com.velocitypowered.api.event.Subscribe;
+import com.velocitypowered.api.event.proxy.ProxyInitializeEvent;
 import com.velocitypowered.api.plugin.Plugin;
 import com.velocitypowered.api.proxy.ProxyServer;
+import dev.httpmarco.polocloud.proxy.config.Config;
+import dev.httpmarco.polocloud.proxy.listener.ServerPostConnectListener;
+import dev.httpmarco.polocloud.proxy.tablist.TablistManager;
+import lombok.Getter;
 
 import javax.inject.Inject;
 
+@Getter
 @Plugin(id = "polocloud-proxy", name = "PoloCloud", version = "1.0.0", authors = "HttpMarco")
 public final class VelocityPlatformPlugin {
 
+    private final ProxyServer server;
+    private Config config;
+    private TablistManager tabManager;
+
     @Inject
     public VelocityPlatformPlugin(ProxyServer server) {
+        this.server = server;
+    }
 
+    @Subscribe
+    public void onProxyInitialize(ProxyInitializeEvent event) {
+        this.config = new Config();
+        this.tabManager = new TablistManager(this);
+
+        var eventManager = this.server.getEventManager();
+        eventManager.register(this, new ServerPostConnectListener(this));
     }
 }
